Use docker-compose for containerizing the application
For local devlopment and testing the container has to be run with
`APP_ENVIRONMENT=local` and the test command must be
`APP_ENVIRONMENT=testing cargo test`.

This is because we need a different host for connecting to the database
for testing than the one used by the server inside the container.